use scopeguard defer to handle drop
